MAINT: move numpy/random/examples -> numpy/random/_examples
