+++ numpy/random/_examples/cython/extending.pyx
