+++ numpy/random/_examples/cython/extending_distributions.pyx
