+++ numpy/random/_examples/cython/setup.py
