+++ numpy/random/_examples/numba/extending.py
